expose kubeObjectDetailRegistry via extension-api
