+++ src/extensions/registries/index.ts
@@ -4,5 +4,6 @@ export * from "./page-registry"
 export * from "./menu-registry"
 export * from "./app-preference-registry"
 export * from "./status-bar-registry"
+export * from "./kube-object-detail-registry";
 export * from "./kube-object-menu-registry";
 export * from "./cluster-feature-registry"
